% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DART_VC-measure.R
\name{DartVC}
\alias{DartVC}
\title{Perform variable selection using DART VC-measure}
\usage{
DartVC(y, X, seed = 123, Lrep = 10, backend = c("dartMachine", "BART"))
}
\arguments{
\item{y}{Response vector.}

\item{X}{Predictor or feature matrix.}

\item{seed}{Random seed.}

\item{Lrep}{Number of replications.}

\item{backend}{The R package used to implement DART. The two choices are \code{"BART"} and \code{"dartMachine"} (the one used in the paper).}
}
\value{
A list containing
\itemize{
\item \code{pos_idx}: selected variable indices
\item \code{Z}: VC-measure summary matrix (\code{\link[=log1p]{log1p()}} transformed follow by \code{\link[=scale]{scale()}})
\item \code{Z_raw}: the raw VC-measure summary matrix
}
}
\description{
This function trains \code{Lrep} (default to 10) DART models using different random
seeds, and extracts the VC-measure from each fit. Hierarchical Agglomerative Clustering (HAC)
is then performed on the VC-measure and identifies important predictors.
}
\examples{
#' # Example code using `BART` R package as the backend
\dontrun{
library(bayesVC)

set.seed(123)
n <- 1000
p <- 100
Lrep <- 10

X <- matrix(runif(n * p), n, p)
y_mu <- 10 * sin(pi * X[, 1] * X[, 2]) + 20 * (X[, 3] - 0.5)^2 + 10 * X[, 4] + 5 * X[, 5]
eps <- rnorm(n, mean = 0, sd = 2)
y <- y_mu + eps

VC_result <- DartVC(
  y = y,
  X = X,
  seed = 123,
  Lrep = Lrep,
  backend = "BART"
)
VC_result$pos_idx
}

# Example code using `dartMachine` R package as the backend
\dontrun{
# Must allocate memory before loading `bayesVC` package when using `dartMachine` backend
# Here I allocated 5GB of memory for Java
options(java.parameters = c("-Xmx5g"))
library(bayesVC)

set.seed(123)
n <- 1000
p <- 100
Lrep <- 10

X <- matrix(runif(n * p), n, p)
y_mu <- 10 * sin(pi * X[, 1] * X[, 2]) + 20 * (X[, 3] - 0.5)^2 + 10 * X[, 4] + 5 * X[, 5]
eps <- rnorm(n, mean = 0, sd = 2)
y <- y_mu + eps

VC_result <- DartVC(
  y = y,
  X = X,
  seed = 123,
  Lrep = Lrep,
  backend = "dartMachine"
)
VC_result$pos_idx
}

}
